fix: still let CSS build prog declare artifact dependency
Initially, I crippled the build prog for the CSS file so much that
it didn't get a chance to declare the dependency of the built CSS
file on the input CSS file.  This resulted in the built CSS being
pruned after a `lektor build`.

Here we cripple the build program just enough that it doesn't
overwrite the output CSS file, but not so much that it doesn't
declare the dependency.