feat: Support dynamic (Jinja-generated) classes (#8)
* Support dynamic (Jinja-generated) classes

Run tailwindcss on the build dir, rather than input templates, so that dynamically-generated values are properly seen by tailwind.

* Don't use Tailwind watcher

Running Tailwind watcher from within Lektor watcher is unnecessary,
since we already have a watcher and know when to rebuild things, and
it leads to bugs, such as the Tailwind watcher leaking memory or
kicking in before the build is fully done.

* Fix tests and README

* Fix timeouts in test_server_*_update()

This was caused by the switch from having two watchers (lektor server and
tailwindcss) to just the server. Because we made writes as soon as
wait_for_server() returns, this would essentially catch the server in
a short window before it was done returning from a build, and it
wouldn't pick up the changes, so the rebuild would never
happen. Previously this was not a concern, because the CSS was watched
by one process, and other files by the other, and they delayed each
other, so there was never a chance to write immediately after the
rebuild was done. Now we sleep for a second before writing to prevent
that.

Note that this is purely an artefact of the testing process, and the
fact we can do it within milliseconds. This is not possible for a
human, so it doesn't come up in normal usage.